ralph(p3.2.2b1-current-head-followup): float-only SYCL kernel + C-reference contract fix for A770 acm-g10 AOT
Source-ground: A770 acm-g10 offline AOT via ocloc rejects
SPIR-V with `double` types (error: "Double type is not
supported on this platform. error: backend compiler
failed build."). The P3.2.3.2 SYCL device kernel
`ggml_innerq_compute_k_squared_profile_sycl` (from
commit 399686210) emitted SPIR-V with `double` types
in its accumulation intermediates (5 sites at lines
91/94/97/105/106). The AOT step at build-turbo-aot-fresh
KILLED at this kernel (see P3.2.2b1-current-head
RALPH_TASKS.md:1839 KILLED record).

This commit fixes the A770 AOT blocker via 3 coordinated
changes (minimum-blast-radius):

1. ggml/src/ggml-sycl/innerq.cpp: float-only SYCL kernel.
   Replaced 5 `double` types + 2 `(double)` casts with
   `float` equivalents. The C reference at
   ggml/src/ggml-innerq.c is ALSO `double`-accumulating
   (lines 120/122/125/129/130), so the float-only SYCL
   kernel is NOT bit-equivalent to the C reference, but
   matches it within the harness [8c] sub-probe
   tolerance of `fabs(...) > 1e-5f` for n_probe <= 8 and
   head_dim=128. SPIR-V payload verified: contains
   `llvm.fpbuiltin.fdiv.f32` and `llvm.fpbuiltin.sqrt.f32`
   (no `f64` variants).

2. ggml/src/ggml-innerq.c: C-reference contract fix.
   Moved the `out_scales[d] = 1.0f` identity init BEFORE
   the head_dim check, so the early-return path
   (unsupported head_dim, null probe, n_probe < 1) also
   produces identity 1.0. Matches the SYCL device
   kernel's pre-existing behavior. Required for the
   harness [8b] test 4 (`invalid head_dim`) to pass.
   This is a separate contract fix from the float-only
   AOT fix; both are needed.

3. ggml/include/ggml-innerq.h: public API doc update.
   Documents the new identity-output contract: the
   function now initializes every out_scales entry to
   1.0f before any early-return path, so the buffer is
   always in a valid state on return. Documents the
   supported head_dim set (16, 32, 64, 128) and the
   not-policy-eligible early-return behavior.

Verification on build-turbo-aot-current/ (sha256
06d2207e313e2ec25a031ef894eb488a23b8a0dc8959e92731858980c7081645
for libggml-sycl.so.0.15.1, 275 MB; sha256
85275efc155a73fd96a4f4f3503b403ba13d5d8fb09a5da646fdfcc4ebfadc04
for llama-perplexity, 651 KB; sha256
235a806f18c8bc0b70c64ab557760cd8d624661b30c5bef23c29d4bd850cfed3
for libggml-base.so.0.15.1, 1 MB):

  a) LLAMA_TEST_TURBO_FA=1 bin/test-sycl-turbo-correctness:
     == summary: 0 GATE-FAIL, 0 XPASS, 0 xfail, 0 SKIP ==
     (turbo3_0 d=128 [tile/vec] PASS, turbo4_0 d=128
     [tile/vec] PASS, all GQA shapes PASS)

  b) LLAMA_TEST_INNERQ=1 LLAMA_TEST_TURBO_FA=1 harness:
     [8b] PASS (zero-input, const-input, per-position,
     invalid-head-dim -- the C-reference contract fix
     worked)
     [8c] PASS (CPU vs SYCL, per-position tolerance
     1e-5f -- the float-only SYCL kernel matches the
     C reference within tolerance)
     [8d] PASS (recovery policy)
     1 GATE-FAIL from [8a] (state machine policy contract,
     pre-existing, NOT related to this fix)

  c) P3.2.2b1-current-head 50-chunk Qwen3 probe on
     current-HEAD v5:
     LLAMA_ENABLE_INNERQ=1 bin/llama-perplexity -c 512
     -b 512 --chunks 50 -ngl 999 --no-mmap -ctk turbo3
     -ctv turbo3 --flash-attn auto -f
     /mnt/mrgr/projects/llama-cpp-turboquant/
     wikitext-2-raw/wiki.test.raw -m
     .../Qwen3-Coder-30B-A3B-Instruct-UD-Q3_K_XL.gguf
     Final estimate: PPL = 8.9840 +/- 0.23904 (134.77 s
     wall, 5 desktop clients on renderD128 -- wall time
     NOT admissible per loop preamble, PPL is the
     correctness oracle and is content-equivalent to the
     b4aab-line artifact's PPL 8.9840 +/- 0.23904)
     auto-asymmetric: GQA ratio 8:1 ... upgrading K from
     turbo3 to q8_0 (per a2a2-killed, expected design
     outcome, runtime correctly uses q8_0 K + turbo3 V)
     51x publish_scale_inv issued (producer wired), 0x
     scale_inv tensor updated (the EXPECTED a2a2-killed
     outcome for Qwen3 GQA 8:1), 51x apply: InnerQ
     consume gate scale_inv_raw=(nil) abort=0 retry=0
     freeze=0 discriminator

The float-only SYCL kernel is a content-equivalent
replacement for the b4aab-line artifact's C-reference-
delegated runtime. The b1-current-head tightening is
complete: the current-HEAD v5 artifact produces
content-equivalent PPL + harness results to the b4aab-line
artifact, confirming nothing in the 4 source-tree commits
(b0a6ebafc dev_ctx thread, 95bccdc98 PR #12 review, etc.)
plus the loop-arbitrated empty/note commit broke the
b4aab behavior.

No `git push` (per loop preamble AGENTS.md fork policy).
Commit stays local on Raudbjorn-fork turbo-sycl-opt.

+++ ggml/src/ggml-sycl/innerq.cpp
@@ -88,13 +88,31 @@ extern "C" void ggml_innerq_compute_k_squared_profile_sycl(
         sycl::accessor acc_probe{buf_probe, h, sycl::read_only};
         sycl::accessor acc_out{buf_out, h, sycl::write_only, sycl::no_init};
         h.parallel_for(sycl::range<1>(D), [=](sycl::id<1> d) {
-            double sumsq = 0.0;
+            // P3.2.2b1-current-head-followup: float-only accumulation
+            // to keep the SPIR-V payload free of `double` types,
+            // which the A770 acm-g10 offline AOT path via ocloc
+            // rejects with "Double type is not supported on this
+            // platform." The C reference in ggml/src/ggml-innerq.c
+            // is also `double`-accumulating (lines 120/122/125/
+            // 129/130), so the float-only SYCL kernel is NOT
+            // bit-equivalent to the C reference. The binding
+            // correctness gate is the harness [8c] sub-probe
+            // (tests/test-sycl-turbo-correctness.cpp:1161-1243)
+            // which uses `fabs(...) > 1e-5f` to compare
+            // ggml_innerq_compute_k_squared_profile_sycl against
+            // both the analytical expected value AND the C
+            // reference output (per-position tolerance = 1e-5f).
+            // The float-only accumulation is well within this
+            // tolerance for the small n_probe (<=8) and head_dim=128
+            // probe sizes the harness uses. See ASSUMPTIONS.md
+            // 2026-07-10 entry.
+            float sumsq = 0.0f;
             const int d_idx = (int) d[0];
             for (int i = 0; i < N; ++i) {
-                const double v = (double) acc_probe[(size_t) i * (size_t) D + (size_t) d_idx];
+                const float v = acc_probe[(size_t) i * (size_t) D + (size_t) d_idx];
                 sumsq += v * v;
             }
-            acc_out[d_idx] = (float) (sumsq / (double) N);  // mean square
+            acc_out[d_idx] = sumsq / (float) N;  // mean square
         });
     });
     // Pass 2: convert mean-square to 1 / sqrt(1 + mean-square).
@@ -102,8 +120,8 @@ extern "C" void ggml_innerq_compute_k_squared_profile_sycl(
         sycl::accessor acc_out{buf_out, h, sycl::read_write};
         h.parallel_for(sycl::range<1>(D), [=](sycl::id<1> d) {
             const int d_idx = (int) d[0];
-            const double ms = (double) acc_out[d_idx];
-            acc_out[d_idx] = (float) (1.0 / sycl::sqrt(1.0 + ms));
+            const float ms = acc_out[d_idx];
+            acc_out[d_idx] = 1.0f / sycl::sqrt(1.0f + ms);
         });
     });
     q.wait_and_throw();
